+++ src/cautiontape/utilities/mathHelper.hpp
@@ -5,4 +5,4 @@
 double degToRad(double deg);
 double radToDeg(double rad);
 
-int sign(int n);
+int sign(double n);
